+++ media/locales/ja/messages.js
@@ -54,7 +54,6 @@ window.languageManager.loadMessages('ja', {
 	CATEGORY_ARDUINO: 'Arduino',
 	CATEGORY_SENSORS: 'センサー',
 	CATEGORY_MOTORS: 'モーター',
-
 	// Servo Block Labels
 	SERVO_SETUP: 'サーボモーターの設定',
 	SERVO_PIN: 'ピン',
@@ -62,6 +61,8 @@ window.languageManager.loadMessages('ja', {
 	SERVO_MOVE: 'サーボモーターを回転',
 	SERVO_ANGLE: '角度',
 	SERVO_MOVE_TOOLTIP: 'サーボモーターを特定の角度に回転させる',
+	SERVO_STOP: 'サーボモーターを停止',
+	SERVO_STOP_TOOLTIP: 'サーボモーターの信号出力を停止する',
 
 	// Arduino Block Labels
 	ARDUINO_SETUP: 'セットアップ',